+++ common/crypto/iss/v1/tests/BUILD
@@ -4,7 +4,7 @@ cc_test(
         "test_iss_curl.c",
     ],
     deps = [
-        "//common/sign/v1:iss_curl",
+        "//common/crypto/iss/v1:iss_curl",
         "//common/trinary:trit_tryte",
         "@unity",
     ],
@@ -16,7 +16,7 @@ cc_test(
         "test_iss_kerl.c",
     ],
     deps = [
-        "//common/sign/v1:iss_kerl",
+        "//common/crypto/iss/v1:iss_kerl",
         "//common/trinary:trit_tryte",
         "@unity",
     ],
