+++ common/crypto/iss/v2/tests/test_iss_curl.c
@@ -9,7 +9,7 @@
 #include <unity/unity.h>
 
 #include "common/crypto/curl-p/trit.h"
-#include "common/sign/v2/iss_curl.h"
+#include "common/crypto/iss/v2/iss_curl.h"
 #include "common/trinary/trit_tryte.h"
 
 #define TEST_SEED                                                              \
