+++ common/helpers/checksum.c
@@ -10,10 +10,10 @@
 #include <stdlib.h>
 #include <string.h>
 
+#include "common/crypto/iss/v1/iss_kerl.h"
 #include "common/crypto/kerl/hash.h"
 #include "common/defs.h"
 #include "common/helpers/checksum.h"
-#include "common/sign/v1/iss_kerl.h"
 #include "common/trinary/trit_tryte.h"
 #include "utils/export.h"
