+++ common/helpers/sign.c
@@ -10,11 +10,11 @@
 #include <stdlib.h>
 #include <string.h>
 
+#include "common/crypto/iss/normalize.h"
+#include "common/crypto/iss/v1/iss_kerl.h"
 #include "common/crypto/kerl/kerl.h"
 #include "common/defs.h"
 #include "common/helpers/sign.h"
-#include "common/sign/normalize.h"
-#include "common/sign/v1/iss_kerl.h"
 #include "common/trinary/trit_tryte.h"
 #include "utils/export.h"
