+++ common/helpers/sign.h
@@ -10,7 +10,6 @@
 
 #include <stddef.h>
 
-#include "common/sign/v1/iss_kerl.h"
 #include "common/trinary/flex_trit.h"
 #include "utils/export.h"
 #include "utils/memset_safe.h"
