+++ consensus/milestone_tracker/milestone_tracker.c
@@ -8,9 +8,9 @@
 #include <inttypes.h>
 #include <stdlib.h>
 
+#include "common/crypto/iss/normalize.h"
+#include "common/crypto/iss/v1/iss_curl.h"
 #include "common/model/milestone.h"
-#include "common/sign/normalize.h"
-#include "common/sign/v1/iss_curl.h"
 #include "common/storage/sql/defs.h"
 #include "common/trinary/trit_array.h"
 #include "common/trinary/trit_long.h"
